squash! fixup! Prevent adding user beyond the Codenvy license
Fix shouldThrowAnExceptionWhenUserBeyondTheLicense test
