+++ cdec/installation-manager-core/src/main/java/com/codenvy/im/managers/Codenvy4xLicenseManager.java
@@ -14,11 +14,11 @@
  */
 package com.codenvy.im.managers;
 
-import com.codenvy.api.license.shared.CodenvyLicense;
-import com.codenvy.api.license.shared.CodenvyLicenseFactory;
-import com.codenvy.api.license.shared.exception.InvalidLicenseException;
-import com.codenvy.api.license.shared.exception.LicenseException;
-import com.codenvy.api.license.shared.exception.LicenseNotFoundException;
+import com.codenvy.api.license.CodenvyLicense;
+import com.codenvy.api.license.CodenvyLicenseFactory;
+import com.codenvy.api.license.exception.InvalidLicenseException;
+import com.codenvy.api.license.exception.LicenseException;
+import com.codenvy.api.license.exception.LicenseNotFoundException;
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
