+++ cdec/installation-manager-core/src/main/java/com/codenvy/im/managers/helper/NodeManagerHelperCodenvy4Impl.java
@@ -14,10 +14,10 @@
  */
 package com.codenvy.im.managers.helper;
 
-import com.codenvy.api.license.shared.CodenvyLicense;
-import com.codenvy.api.license.shared.exception.InvalidLicenseException;
-import com.codenvy.api.license.shared.exception.LicenseException;
-import com.codenvy.api.license.shared.exception.LicenseNotFoundException;
+import com.codenvy.api.license.CodenvyLicense;
+import com.codenvy.api.license.exception.InvalidLicenseException;
+import com.codenvy.api.license.exception.LicenseException;
+import com.codenvy.api.license.exception.LicenseNotFoundException;
 import com.codenvy.im.artifacts.CDECArtifact;
 import com.codenvy.im.artifacts.helper.SystemProxySettings;
 import com.codenvy.im.commands.Command;
